Skeleton for Player Entry Screen
Work in progress for the player entry screen. It isn't all there and you cannot add players to the database just yet but it is at least an entry screen. We will need to add some functionality to connect the splash screen window and this current. The edit players button only locks the blue player side.