Finished GUI, need add player function
Finished adding all the buttons and setting them correctly in code, fixed issues. The edit players button will lock/unlock the text boxes to be typed in. All that needs to be added is a function to create a player and add them to the table in our database. The database is automatically connected when the program runs. It output the players in the database already to show that it is connected.